+++ src/lib.rs
@@ -1,3 +1,6 @@
+#[global_allocator]
+static ALLOC: mimalloc::MiMalloc = mimalloc::MiMalloc;
+
 pub mod converter;
 pub mod renderer;
 pub mod skip;